machinery v0.2.0 Machinery

Main Machinery module. It keeps the bunk of the Machinery logics, it’s the module that will be imported with use on the module that the state machine will be implemented.

Parameters

opts: A Keyword including states and transitions.

Example

  defmodule Project.User do
    use Machinery,
      states: [:created, :partial, :complete],
      transitions: [
        %{from: :created, to: [:partial, :complete]},
        %{from: :partal, to: :complete}
      ]
  end

Functions

__using__(opts) (macro)

Main macro function that will be executed upon the load of the module using it.

It basically stores the states and transition